Added rate limiting

+++ scripts/reset.sh
@@ -22,14 +22,23 @@ echo "Removing '.nyc_output'..."
 rm -rf .nyc_output
 echo "Removing 'coverage'..."
 rm -rf coverage
+echo "Removing 'dist'..."
+rm -rf dist
 echo "Removing generated wings files..."
 rm -rf src/cockpit/wings/*.ts src/turbine/wings/*.go
-echo "Dropping database"
-psql --username="$USER" -w -c "DROP DATABASE globetrotte;" || echo "Database does not exist. Skipping..."
+
+if [ $1 = "--all" ]; then
+  echo "Dropping database"
+  psql --username="$USER" -w -c "DROP DATABASE globetrotte;" || echo "Database does not exist. Skipping..."
+fi
+
 echo "All removal completed successfully!"
 echo
 echo "Initializing pnpm..."
 pnpm i
-echo
-echo "Creating database..."
-psql --username="$USER" -w -c "CREATE DATABASE globetrotte;"
+
+if [ $1 = "--all" ]; then
+  echo
+  echo "Creating database..."
+  psql --username="$USER" -w -c "CREATE DATABASE globetrotte;"
+fi
